+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 - Added sequential mode in class `Stage`. Use can turn in on by setting `parallel` to *true*.
 - Added external data flow description in pipeline description
 - Added method `beforeAll` into `ConfigLoader`
+- Added new method `addStage` and `addFactory` that take a class object as input. The instantiation will be handled 
+  by the stage. The Factory class must have an empty primary constructor or a primary constructor with only a SparkSession 
+  object.
 
 ## 0.3.0 (2019-07-22)